Add files via upload
Just for reference you have to run each of these files individually first. I would not run the MCTVGui first I would do it like so
MctGui.ps1 >> MctVehicle.ps1 >> MctGui.psi. More than likely this is going to fail your going to get a error you have to refresh the powershell so that it realizes that it is class driven. One of the main reasons I moved everything over to C++, that and speed